Fix abi decoding of bytes32
Note that any indexed string or array will be stored as a hash, so
we need to make that clear. The new Hashed field marks that.

We might want to store strings in bytes32 so that it does not get
hashed. If this is the case then use code similar to the test
TestEventEmitterBytes32isString.

+++ execution/solidity/event_emitter.sol
@@ -6,7 +6,7 @@ contract EventEmitter {
         bytes32 indexed direction,
         bool trueism,
         string german ,
-        int indexed newDepth,
+        int64 indexed newDepth,
         string indexed hash)
         anonymous;
